requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=5.4.0
tqdm>=4.62.0
python-dotenv>=0.19.0

# API clients
openai>=1.0.0
anthropic>=0.18.0
cohere>=4.0.0

# Evaluation and metrics
scikit-learn>=1.0.0
nltk>=3.7
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.96

# Data processing
jsonlines>=3.0.0
datasets>=2.10.0
langdetect>=1.0.9
polyglot>=16.7.4

# Visualization and reporting
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
jinja2>=3.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-asyncio>=0.18.0

# Development tools
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
pre-commit>=2.17.0

# Logging and monitoring
loguru>=0.6.0
rich>=12.0.0

# Optional dependencies for advanced features
# Uncomment if needed:
# tensorflow>=2.11.0  # For TensorFlow-based models
# spacy>=3.4.0  # For advanced NLP processing
# fastapi>=0.95.0  # For API server
# uvicorn>=0.20.0  # For API server